Improvements related to file watch tools

+++ ox_mon/triggers/file_triggers.py
@@ -1,6 +1,7 @@
 """Module for trigger on file related events.
 """
 
+import time
 import tempfile
 import shutil
 import hashlib
@@ -24,6 +25,10 @@ except Exception as problem:  # pylint: disable=broad-except
 class FileWatchCopy(interface.OxMonTask):
     """Trigger to copy files when they are created or modified.
     """
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.max_loops = None
 
     @classmethod
     def options(cls):
@@ -51,36 +56,72 @@ class FileWatchCopy(interface.OxMonTask):
                 raise ValueError('%s %s must be a directory' % (
                     name, key))
 
-        inote = inotify.adapters.Inotify()
-
-        inote.add_watch(self.config.watch, (
+        inote = inotify.adapters.InotifyTree(self.config.watch, (
             inotify.constants.IN_CLOSE_WRITE |
+            inotify.constants.IN_CREATE |
             inotify.constants.IN_MOVED_TO |
             inotify.constants.IN_MOVE_SELF))
 
         for event in inote.event_gen(yield_nones=False):
             (header, type_names, path, filename) = event
+            self._handle_event(header, type_names, path, filename)
+            if self.max_loops is not None:
+                self.max_loops -= 1
+                if self.max_loops >= 0:
+                    logging.info('%i loops left', self.max_loops)
+                else:
+                    logging.warning('Stopping loop')
+                    break
 
-            logging.debug('HEADER=%s', str(header))
-            logging.info("PATH=[%s] FILENAME=[%s] EVENT_TYPES=%s",
-                         path, filename, type_names)
-            my_fname = os.path.join(path, filename)
-            my_fd, tname = tempfile.mkstemp()
-            os.close(my_fd)
-            logging.info('Copying %s to temp file %s', my_fname, tname)
-            shutil.copy(my_fname, tname)  # copy to tmp file first for safety
-            hash_dir, my_utc = self.prep_destination(tname)
-            if os.path.exists(os.path.join(hash_dir, 'main.data')):
-                logging.info('main data already exists in dir %s', hash_dir)
+    def _handle_event(self, header, type_names, path, filename):
+        logging.debug('HEADER=%s', str(header))
+        logging.info("PATH=[%s] FILENAME=[%s] EVENT_TYPES=%s",
+                     path, filename, type_names)
+        my_fname = os.path.join(path, filename)
+        if self._handled_delete(my_fname, type_names):
+            return
+        if os.path.isdir(my_fname):
+            logging.info('Skip directory %s', my_fname)
+            return
+        my_fd, tname = tempfile.mkstemp()
+        os.close(my_fd)
+        self._do_copy(my_fname, tname)
+
+    @staticmethod
+    def _handled_delete(my_fname: str, type_names) -> bool:
+        result = False
+        if type_names == ['IN_DELETE']:  # only deleting file
+            for stime in [1, 2, 4]:
+                if not os.path.exists(my_fname):
+                    logging.info('Skip %s since it was deleted', my_fname)
+                    result = True
+                    break
+                logging.debug('sleep %s to wait for %s deletion',
+                              stime, my_fname)
+                time.sleep(stime)
             else:
-                logging.info('Copying %s to %s', tname,
-                             os.path.join(hash_dir, 'main.data'))
-                shutil.copy(tname, os.path.join(hash_dir, 'main.data'))
+                logging.error('Got delete for %s but it was not deleted',
+                              my_fname)
+        return result
 
-            with open(os.path.join(hash_dir, 'names.txt'), 'a') as my_fd:
-                my_fd.write('%s: %s\n' % (my_utc, my_fname))
-                logging.info('At utc %s saved %s', my_utc, my_fname)
-            os.remove(tname)
+    def _do_copy(self, my_fname, tname):
+        logging.info('Copying %s to temp file %s', my_fname, tname)
+        if not os.path.exists(my_fname):
+            logging.error('Could not copy %s since it is gone!', my_fname)
+            return
+        shutil.copy(my_fname, tname)  # copy to tmp file first for safety
+        hash_dir, my_utc = self.prep_destination(tname)
+        if os.path.exists(os.path.join(hash_dir, 'main.data')):
+            logging.info('main data already exists in dir %s', hash_dir)
+        else:
+            logging.info('Copying %s to %s', tname,
+                         os.path.join(hash_dir, 'main.data'))
+            shutil.copy(tname, os.path.join(hash_dir, 'main.data'))
+
+        with open(os.path.join(hash_dir, 'names.txt'), 'a') as my_fd:
+            my_fd.write('%s: %s\n' % (my_utc, my_fname))
+            logging.info('At utc %s saved %s', my_utc, my_fname)
+        os.remove(tname)
 
     def prep_destination(self, tname: str):
         """Make destination to store file and return hash and utc timestamp.
@@ -101,8 +142,11 @@ class FileWatchCopy(interface.OxMonTask):
         """
         my_utc, my_hash = self.make_utc_and_hash(tname)
         hash_dir = os.path.join(self.config.archive, my_hash)
-        if not os.path.exists(hash_dir):
+        try:
             os.mkdir(hash_dir)
+        except FileExistsError as prob:
+            logging.info('Ignoring %s', str(prob))
+        assert os.path.exists(hash_dir)
         return hash_dir, my_utc
 
     @staticmethod
